Add support for preferences files using the "plist" DTD.  Abstract the preferences interface (IPreferences) to allow support for plist format, straight XML (not yet implemented) or other formats.  The interface is almost exactly one-to-one with the SmartML Get and Set calls.
Added an interface to allow for a control or class to register a callback for changes to preferences(IpreferencesReceiver).  An example of where this could be applied is the PrintPreview window, where a callback could alert it to changes in the page setup so it could redraw with the new settings.

The user preferences previously set in Globals now reside in a plist Preferences file stored in the PreferencesFolder hierarchy.  Globals can be used by the system admin to set defaults for new users.

Updated MainPreferencesWindow to use the new preferences file.  Incorporated Joshua's changes to show the current document folder.  Rearranged the dialog layout to give more space to the document folder group.  Expanded the width of some controls to avoid truncation of labels for Deutsch/DeutschNeutral and Svenska (will probably help other Germanic/Slavic languages).