Update version to baselevel 16.
Resize MainSettingsWindow.nte_documents_warning (merge of r146 from Joshua's branch).

PresentWindow's Visible attribute set to False to hold display of the window until the last possible moment of the Present method.  Rationale: try to reduce number of visible window updates.

Make PreviewPicture and LastPicture in PresentWindow match CurrentPicture when the window is made visible. This appears to reduce or eliminate the white flash at the start of a presentation when transitions are enabled. (Bugs 1559351 and 1658099)