Correct issue in SEditFieldNumeric where the min/max would be enforced regardless of the EnforceRange setting.
Add conditional code to SButton to fire the Popup in the MouseDown on Macintosh and in MouseUp on Windows and Linux.  This brings it into conformance with the standards for each platform and also makes Popup fire correctly when OpenSong is compiled with RB > 5.5.

Add conditional code to eliminate a SetFocus in PresentWindow.Activate that kept the window from correctly detecting keystrokes when compiled under RB2007r2.