Add call to ClearRect at start of the Repaint. For some reason, the later versions of Xojo needs this (and it should not, but that's Xojo...).
If the font size is set to 0, Xojo uses the system default size. This is typically also used with the Xojo-specific font names "System" and "SmallSystem" to select the default GUI font for the OS. Some of the decorations in SButton would change the font size, and that does not work correctly for font size 0.