Cleaned up some GUI issues...
The enabled controls at program launch or other times when a Song is not selected (e.g., changing song folder and current song is not in the new folder) were inconsistent and actions that should not be available were enabled. This was due in part to the Enabled property in the inspector for the control on the window designer being set to Enabled instead of Disabled and some controls not appearing in the Window's UpdateMenuItems event. Also, the SButton subclass was forcing Enabled to True in the Open event and completely overriding the setting in the window designer.

Also cleaned up enabled/disabled setting for the button to remove linked songs in the song editor.